move a test to a different module
