+++ packages/fuel-indexer-tests/tests/indexing_postgres.rs
@@ -1,7 +1,4 @@
 use bigdecimal::ToPrimitive;
-use fuel_indexer_database::queries::last_block_height_for_indexer;
-use fuel_indexer_database::IndexerConnection;
-use fuel_indexer_lib::config::IndexerConfig;
 use fuel_indexer_tests::fixtures::{
     mock_request, setup_indexing_test_components, IndexingTestComponents,
 };
@@ -597,104 +594,4 @@ async fn test_index_sway_enums() {
 
     assert_eq!(row.get::<&str, usize>(1), EXPECTED_CONTRACT_ID);
     assert_eq!(row.get::<BigDecimal, usize>(2).to_u64().unwrap(), 0);
-}
-
-#[actix_web::test]
-async fn test_replacing_an_indexer_and_keeping_or_removing_data() {
-    let config = IndexerConfig {
-        replace_indexer: true,
-        ..IndexerConfig::default()
-    };
-
-    let IndexingTestComponents {
-        node,
-        mut service,
-        mut manifest,
-        db,
-        ..
-    } = setup_indexing_test_components(Some(config)).await;
-
-    node.abort();
-
-    let mut conn =
-        IndexerConnection::Postgres(Box::new(db.pool.acquire().await.unwrap()));
-
-    let last = last_block_height_for_indexer(
-        &mut conn,
-        manifest.namespace(),
-        manifest.identifier(),
-    )
-    .await
-    .unwrap();
-
-    // Starting from block 1.
-    assert_eq!(last, 1);
-
-    // Add three more blocks.
-    mock_request("/block").await;
-    mock_request("/block").await;
-    mock_request("/block").await;
-
-    let last = last_block_height_for_indexer(
-        &mut conn,
-        manifest.namespace(),
-        manifest.identifier(),
-    )
-    .await
-    .unwrap();
-
-    assert_eq!(last, 4);
-
-    // Set the start block and the end block to 5. This way, since blocks 1-4
-    // are already indexed, this new indexer should do nothing.
-    manifest.set_start_block(5);
-
-    // Replace the indexer, keep the data.
-    service
-        .register_indexer_from_manifest(manifest.clone(), false)
-        .await
-        .unwrap();
-
-    let last = last_block_height_for_indexer(
-        &mut conn,
-        manifest.namespace(),
-        manifest.identifier(),
-    )
-    .await
-    .unwrap();
-
-    // We kept the data, i.e. blocks 1-4, and the redeployed indexer should do
-    // nothing. The last block must still be 4.
-    assert_eq!(last, 4);
-
-    // Replace the indexer, remove the data.
-    service
-        .register_indexer_from_manifest(manifest.clone(), true)
-        .await
-        .unwrap();
-
-    let last = last_block_height_for_indexer(
-        &mut conn,
-        manifest.namespace(),
-        manifest.identifier(),
-    )
-    .await
-    .unwrap();
-
-    // We removed the data, and the new indexer sould not start until block 5 is
-    // present. The last block must be 1.
-    assert_eq!(last, 1);
-
-    // Increase blocks to 5. The new indexer should start.
-    mock_request("/block").await;
-
-    let last = last_block_height_for_indexer(
-        &mut conn,
-        manifest.namespace(),
-        manifest.identifier(),
-    )
-    .await
-    .unwrap();
-
-    assert_eq!(last, 5);
 }
